refactor(session): keep shared protocol at session root
