+++ src/session/agent/commands.test.ts
@@ -14,7 +14,7 @@ import {
   type HunkDaemonCliClient,
 } from "./commands";
 import { HUNK_DAEMON_UPGRADE_RESTART_NOTICE } from "./capabilities";
-import { HUNK_SESSION_API_VERSION, HUNK_SESSION_DAEMON_VERSION } from "./protocol";
+import { HUNK_SESSION_API_VERSION, HUNK_SESSION_DAEMON_VERSION } from "../protocol";
 
 function createTestListedSession(sessionId: string) {
   return buildTestListedSession({
