+++ src/session/agent/commands.ts
@@ -32,7 +32,7 @@ import {
   type HunkSessionCliClient,
 } from "./cliClient";
 import { reportHunkDaemonUpgradeRestart } from "./capabilities";
-import { HUNK_SESSION_API_VERSION, type SessionDaemonAction } from "./protocol";
+import { HUNK_SESSION_API_VERSION, type SessionDaemonAction } from "../protocol";
 
 const REQUIRED_ACTION_BY_COMMAND: Record<SessionCommandInput["action"], SessionDaemonAction> = {
   list: "list",
