+++ src/session/protocol.ts
@@ -8,7 +8,7 @@ import type {
   SessionReloadCommandInput,
   SessionReviewCommandInput,
   SessionSelectorInput,
-} from "../../core/types";
+} from "../core/types";
 import type {
   AppliedCommentBatchResult,
   AppliedCommentResult,
@@ -21,7 +21,7 @@ import type {
   SessionLiveCommentSummary,
   SessionReview,
   SessionReviewNoteSummary,
-} from "../types";
+} from "./types";
 
 export const HUNK_SESSION_API_PATH = "/session-api";
 export const HUNK_SESSION_CAPABILITIES_PATH = `${HUNK_SESSION_API_PATH}/capabilities`;
