+++ src/session/protocolSchemas.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import type { CliInput } from "../../core/types";
+import type { CliInput } from "../core/types";
 import type { SessionDaemonRequest } from "./protocol";
 
 /**
